EXフォルダの作成と一部コードの修正

+++ 第4回/練習問題.py
@@ -1,3 +1,3 @@
-r = int(input('半径を入力:'))
+r = float(input('半径を入力:'))
 PI = 3.14
 print(PI * r ** 2)
